Moves all stories next to their components
